Initial commit.
This includes a scalar wave equation propagator and tests to ensure that
the forward modeled results are within tolerances of theoretical
expectations for constant wave speed and point scatterer models, and
that the model gradient result is also close to the expected.

It does not include any convenience functions such as an FWI
implementation (although the README does link to an example of using it
for FWI), wavefield compression, SEG-Y I/O, anisotropy, or GPU
implementations.