Replaces CFFI to access compiled
Replaces CFFI with PyTorch's recommended extension method
(using C++ and pybind11). Tests pass - results look the same
as with CFFI. Performance also appears to be similar,
but compiling takes longer and requires more memory.

This also requires changes to .travis.yml as 4.9 is
now the minimum supported GCC version, but Travis-CI images
come with a maximum of 4.8, so we now have to update
GCC on Travis-CI (to 4.9) before running pytest.