Adds GPU scalar wave propagators
When PyTorch detects that a compatible GPU is available
(torch.cuda.is_available() == True), GPU implementations of the scalar
wave propagators, written in CUDA, are compiled in addition to the
normal CPU ones. They will run if the model Tensor is on the GPU. The
other Tensors, with the exception of dx, must then also be on the GPU.

To reduce code duplication, scalar.cpp has been split into a part
containing functions that are the same for both CPU and GPU
implementations, and a device-specific part (scalar_cpu.cpp for the CPU
and scalar_gpu.cu for the GPU).