Increases dt in model_grad_2d test
This test took a while to run (about 27 seconds on the machine I tested
it on), most of it calculating the expected value rather than the actual
one. Increasing the time step dt reduced the runtime considerably (to
around 2 seconds). Some other tests still take around 20 seconds to run,
but there isn't an obvious way to reduce those.