Add survey-pad feature.
It is often unnecessary to perform wave simulation on the entire
model. This feature allows users to specify the padding around
the sources and receivers in the shots being propagated that
should be included in the domain used for wave simulation. This
can reduce computational cost when the model is much larger
than the area covered by the shots being propagated.

Closes #5