Adds ability to specify PML width as Tensor.
Previously PML width could only be specified as an int, and
so the same value was used for all dimensions. Now a different
width can be supplied for the beginning and end of each dimension.
This allows a reflective free surface by setting the PML width
for the beginning of the depth dimension to 0, for example.