Fixes bugs with padding and GPU propagator
Padding:

PyTorch's padding utility expects that the padding is specified in
reverse order, but it was not reversed.

Uses atomicAdd for adding sources on GPU:

Previously sources were added directly to the wavefield in global
memory, but if two sources were in the same grid cell, then there would
be a race condition. Using atomicAdd ensures that all sources get added
to the wavefield correctly, but may reduce performance.

GPU Tensors:

Previously there were some missing checks on whether a Tensor was on the
GPU or CPU.