Improves GPU performance
Two changes are made to improve the performance of the GPU propagators:

* The finite difference coefficients are put into constant memory
* The PML calculation is only done in the regions where it is needed
(the PML region and PAD cells inside it), while in the interior the
simple scalar wave equation is used instead.

Using shared memory for the spatial finite difference calculations did
not seem to improve performance, possibly because the results are
already cached. There seemed to be a small performance improvement by
using shuffle operations to calculate the spatial finite differences,
but the additional code complexity did not justify this optimization.

Applying the imaging condition and adding the sources consume negligible
runtime compared to the wave equation calculation, and so code
modifications to avoid atomicAdd were not justified.

Previous experience with cuDNN suggested that it would not improve the
performance of the spatial finite difference calculations, but this may
need to be revisited in the future.

No threadblock size tuning was done. This should be revisited in the
future.

Closes #15