Replaces flip function with new built-in version
Some of the tests in test_scalar needed to reverse the axis of a Tensor
(such as when reversing time). As PyTorch does not allow negative Tensor
strides, I had to create a function to flip the appropriate axis.
PyTorch 0.4.1 adds a new flip method, so I now use that instead.