Separates pad and extract, and allows double prec.
Previously padding the model with cells to make spatial finite
difference cleaner, and for the PML, was done inside scalar.py. They are
now separate differentiable modules that are chained together before
calling the propagator. This is more elegant, but it also improves the
accuracy of backpropagation as the effect of the padded area is now
included (an admittedly small effect). The cost of this is that the
imaging condition in the propagator now has to be applied to the whole
domain (i.e., including padding) rather than just the interior -
somewhat increasing memory usage and computation and also requiring
changes to the imaging condition so that it is still correct in the PML
regions. New tests accompany the additional modules.

The other major change is that the CPU propagator is now compiled for
both single and double precision. Double precision is unlikely to be of
interest for seismic propagation, but it is useful for testing the
propagator. PyTorch's gradcheck is now used to verify the propagator
instead of custom tests previously used. Fully accounting for the effect
of the PML in the 3D propagator would be expensive as it would require
storing additional wavefields, so the expensive terms are omitted. As a
result, the 3D gradient is not exactly right, and so the accuracy limit
of gradcheck has to be increased for that case. The gradcheck tests also
fail for the GPU propagator (at least in 2D and 3D) due to them not
being reentrant. I believe this is because functions such as atomicAdd
are not deterministic. When I checked, the difference between repeated
calls to the 2D propagator was O(1e-10), so is in reality insignificant.
Double precision is not enabled for the GPU as atomicAdd of doubles only
works on newer GPUs. This could be enabled by editing setup.py.