Remove the INPLACE strategy as unlikely to be used
Previously there were two strategies for saving the forward wavefield
(so that it could be recalled during backpropagation): INPLACE and COPY.
The INPLACE strategy was for when the finite difference and source
wavelet time step intervals were the same (the number of inner steps was
1). It used the same array in memory for propagation and storage. The
COPY strategy instead copied the time steps of the forward wavefield
that need to be recalled to a separate memory location. As the finite
difference time step interval is likely to always be smaller than the
imaging condition interval (the Nyquist rate of the seismic data), the
INPLACE strategy is unlikely to be used in practice, and so this commit
removes it.