Resample source amplitude to inner step length
Previously "nearest" interpolation was used for the source amplitude.
This was not very accurate if the source amplitude sampling was
significantly coarser than the finite difference (inner step) time
sampling. The interpolation is now done using SciPy's resampling
function. This uses more memory but should be a bit more accurate.

Currently the outer time step sampling should still be quite fine as the
time derivative of the saved wavefields will be inaccurate otherwise,
and there is also a possibility that some things (such as the receiver
amplitude) are sampled one time step off from what they should be. This
will be resolved in a future commit.