Tidies and improves accuracy of compiled code
This commit removes the preprocessor macro functions that were used to
implement derivatives in the C++ and CUDA code. They are replaced by
regular functions. This is intended to make the code easier to
understand. The code has also been tidied, in particular, rather than
using multiple #ifdefs within functions such as the imaging condition to
select the behaviour depending on the dimension of the model, separate
functions for the different dimensions are now used. This makes the code
longer and causes repetition, but hopefully makes the code clearer.

Another major change in this commit is that the first and second time
derivatives of the wavefield are now calculated and saved during the
forward pass. This uses more memory (the stored wavefields array is now
three times the size), but should give more accurate results when the
time interval between saved wavefields is large.

The first sample in the recorded receiver data now corresponds to time
zero, instead of being one time step into the simulation as before. This
seems reasonable, but seemed to require that some error
tolerances in the test suited be slightly less stringent.