Switch to PyTorch v1.0.0
PyTorch 1.0.0 is now a stable release, so this commit switches from
0.4.1 to 1.0.0. No changes were required, but to avoid warning messages
the torch.h header file has been replaced by extension.h, and in
model.py there is a test to check if a value is already a Tensor before
converting it. The test suite seems to take longer to run on Travis than
it did with PyTorch 0.4.1, which may indicate performance regressions
(15 minutes vs 11 minutes).