Remove __restrict and default(none)
The qualifier __restrict tells the compiler that there is no
aliasing, which may allow it to perform better optimizations,
but it is a non-standard extension that might cause trouble
for some users, so I have removed it.

Using default(none) in OpenMP pragmas is generally considered
good practice, as it avoids accidentally sharing variables. I
previously used it, but doing this causes problems for some
users as some compilers automatically share const variables
while others do not, and explicitly sharing a variable that
has already been automatically shared causes a compiler error.
Removing default(none) avoids this problem.

I also changed from .data<> to .data_ptr<>, as the former is
now apparently deprecated.