Adds Nix tests
This commit, contributed by @AtilaSaraiva, contains a GitHub Action
that builds Deepwave using Nix and runs Deepwave's tests.

@AtilaSaraiva maintains the Nix package for Deepwave, and found that
one of Deepwave's test was failing in the last release when built
with Nix, even though it passed when built the regular way. Adding
this new Action ensures that future releases will also pass all
of the tests when built using Nix.

In order for the problematic test
(test_born_gradcheck_2d_different_pml)
to pass with Nix, this commit increases the tolerance used from the
default 1e-8 to 5e-8.

Closes #45
Closes #46