Do not zero edges of v, b, and a
Since v0.0.13 Deepwave has zeroed the edges of the velocity model
and the a and b arrays (for the PML). This commit instead switches
to zeroing the edges of the wavefields. The reason for the change
is that the previous approach caused instability that resulted in
the wavefield exploding to high amplitudes.