Minor bugfixes for PyTorch v1.9
In PyTorch v1.9, torch.real can only be applied to complex
numbers (it raises an error if applied to a real number), and
torch.pi does not exist. I fix both of these (letting the cosine
taper be complex if the input is complex, and using math.pi) so
that all of the code still runs on PyTorch v1.9 (the oldest
version supported by Deepwave). Thanks to @guoketing for reporting.