Add GSOT and JMI examples
Graph Space Optimal Transport is a recently proposed loss function
for seismic inversion that reduces the problem of cycle-skipping.
This commit adds a simple example demonstrating it.

The new ability to record from receivers in the background wavefield
of the Born propagator is used in the second new example, which
demonstrates performing joint migration inversion.