Use tuples in Hicks dictionaries
The Hicks location interpolation code previously used Tensors
as the keys in the dictionaries, so that it would be possible
to apply PyTorch's JIT to the code (it does not support tuples
as dictionary keys). However, I have discovered that Tensors
are considered different even if their contents is the same,
so the dictionaries were not matching duplicates as I hoped.
This commit changes to use tuples for the keys, even though it
will now not be possible to use JIT.