Switch to distributing precompiled libraries
Previous versions of Deepwave were distributed purely as code. The
C/C++/CUDA code was then compiled on the user's system. This had
advantages, such as the compiled code being optimised for the
user's system, but had the big disadvantage of some users
(especially those not using Linux) having difficulty with this
compilation step. This is a big commit that changes distribution
so code is instead compiled, using GitHub Actions, into
shared/dynamic libraries. Most users should now not have to
compile code, but those who want to, such as those with
unusual systems that cannot run the precompiled version,
can still build from source.

This commit also makes a small change to the scalar_born propagator
so that it handles scatterers at source locations more accurately,
and includes documentation improvements such as a new section on
what Deepwave calculates and a reorganisation of examples. It
reverts a previously committed but not released change of adding
pressure sources to the elastic propagator.