Add time_taper option to apply Hann window
Deepwave samples the integration over timesteps to calculate
gradients with respect to model parameters at a frequency that
relies on signals being bandlimited. When signals do not start
from zero amplitude, high frequencies are introduced which can
break this assumption. To ensure that this does not occur, this
commit adds a new time_taper option that applies a Hann window
to source and receiver amplitudes. It is off by default, and is
probably mainly useful during testing of the propagators.