Change build system and tidy code
Switch from build scripts for each operating system and distributing
sdists with the compiled libraries included, to using
cibuildwheel and scikit-build-core. The precompiled libraries are
now distributed in wheels. This is a more standard approach, reduces
the file size (by only including the compiled library of the target
system, rather than all of them as before), and means that it is
easier for users to build from source if desired. [ci skip]