Add callbacks feature and examples.
This commit adds new parameters forward_callback, backward_callback,
and callback_frequency to the propagators. These enable users to
run arbitrary Python code after each forward and backward time step,
with access to the wavefields, models, and gradients.

Although this does not enable anything that was not possible before,
as it was previously possible to upsample the source amplitudes to
the CFL time interval and then propagate one step at a time, the
new callback approach is much simpler (avoiding the need to
upsample the sources and downsample the receivers, and repeatedly
call the propagator with different chunks of the source amplitudes),
and also much faster (by avoiding having to redo a lot of propagator
setup that is done every time the propagator is called).

This commit also adds a new example that demonstrates using callbacks
to record forward and backward wavefields and the gradient. The
snapshot component of the target final wavefield example is also
changed to use callbacks, and the custom imaging condition example
is completely rewritten to use callbacks to implement a simple
illumination compensation.