Switch CUDA_ARCHITECTURES from all-major to all
With the switch to CMake in v0.0.21, the list of CUDA architectures
that are compiled for was changed from a manual list of many
architectures to CMake's "all-major" option. This commit changes
to CMake's "all" option, as I believe the small increase in file
size is worth the benefit of having compilation specialised for
more architectures. This change seems to resolve a problem where
certain kernels with high register requirements were failing on
a GPU with a compute capability that was not compiled for.