Adds high step_ratio warning
Deepwave may internally use a smaller time step interval than the
one provided by the user, to obey the CFL condition. If the internal
one is much smaller than the one provided by the user, it might
indicate that something is wrong. This commit adds a warning when
the ratio between these (referred to internally as `step_ratio`)
is 20 or higher.