Switch W-AFDA to Zeng and add Python backend.
Deepwave's elastic propagator previously used the W-AFDA method of
Kristek et al. to implement free surfaces. This performed well, but had
some limitations, including that only planar free surfaces along the
edges were supported, and only 2nd and 4th order accurate spatial finite
differences were allowed. The code was also very complicated, making it
hard to create a Python backend implementation with reasonable
performance and making a 3D implementation unappealing.

This commit switches to instead use the Improved Vacuum Method of
Zeng et al. (2012). The code is substantially simpler, it enables
topography, and it works with higher order spatial finite differences,
although achieving comparable accuracy to W-AFDA on a planar free
surface may require a higher number of grid cells per wavelength.

This simplification of the elastic propagator also enables the commit to
add a Python backend for the propagators. This makes testing easier
(outputs can be compared with the Python implementation, which is
especially useful for testing the gradients as it avoids the very
time-consuming torch.autograd.gradcheck), and means that Deepwave can be
used on devices that are supported by PyTorch but not by Deepwave's
compiled code.

[ci skip]