Add elastic pressure sources.
Previously Deepwave only had body force sources in the two spatial
dimensions. As demonstrated in the documentation, it is possible to
generate pressure sources using the body force sources. It is, however,
somewhat cumbersome, so, given how common pressure sources are, this
commit adds the ability to specify pressure sources directly.