Adds intermediate data storage options.
Intermediate data calculated during the forward pass (commonly thought
of as snapshots of the wavefields, although in Deepwave they are not
exactly the wavefields) needs to be saved for backpropagation in order
to calculate gradients with respect to property models. This can
require a substantial amount of memory, especially for 3D propagation.
Deepwave previously stored all of this data uncompressed in the memory
of the compute device. This commit adds the option to apply lossy
compression to the data (a simple compression to 8 bits per element) and
to store the data in compute device memory, CPU memory, or on disk. It
also provides the option to prevent Deepwave from storing this
intermediate data and calculating gradients with respect to the property
models, which avoids unnecessary memory consumption and computation for
users who use the callback facility to handle this themselves.
[ci skip]