Uses double buffering to improve CUDA storage speed.
The propagators need to store intermediate data during forward
propagation to calculate gradients with respect to input property models
during backpropagation. Deepwave recently added the ability to control
this storage: where it is stored and whether compression is applied. All
operations were performed sequentially, however, and so this could waste
GPU resources by leaving the GPU idle while data was being transferred
to the CPU, for example.

This commit uses CUDA streams and double buffering to overlap
computation and storage in certain circumstances, reducing the runtime
overhead of storage. It does not do this when the data is to stored on
disk, however, as that requires a blocking OS call that prevents
overlap. This could be resolved by using threads. [ci skip]